Modified the script to apply idle and walking animations
Removed some of the original variables (AnimationPlayer and Sprite2D) to replace it with the new AnimatedSprite2D node as an onready variable.

Then applied the idle and walk animations in the match_movement_animation and match_idle functions, under the animation region

Movement animations do not trigger properly when playtesting however, as the character sprite remains stuck to the idle down animation. To be addressed at a later date.